Merge continuous collinear wall segments.
Previously we'd end up with one of the wall runs split in the middle.
This is something that couldn't be resolved prior to wall planning because
only then are corners translated into distinct walls.  (That is, this can't
be handled by changing the startFace since what we really need to be doing
is splitting a face in its middle.)

This will likely need to be changed to consider the line tangents at the
join point in the future in order to support curved corners, etc.